chore: move utils back to /lib
